# WUI based implementation to initUndistortRectifyMap