--- a/modules/imgproc/src/undistort.cpp
+++ b/modules/imgproc/src/undistort.cpp
@@ -43,6 +43,7 @@
 #include "precomp.hpp"
 #include "opencv2/imgproc/detail/distortion_model.hpp"
 #include "undistort.hpp"
+#include "opencv2/core/hal/intrin.hpp"
 
 namespace cv
 {
@@ -102,6 +103,14 @@ public:
 #if CV_TRY_AVX2
         useAVX2 = checkHardwareSupport(CV_CPU_AVX2);
 #endif
+#if CV_SIMD_64F
+        for (int i = 0; i < 2 * v_float64::nlanes; ++i)
+        {
+            s_x[i] = ir[0] * i;
+            s_y[i] = ir[3] * i;
+            s_w[i] = ir[6] * i;
+        }
+#endif
     }
 
     void operator()( const cv::Range& range ) const CV_OVERRIDE
@@ -126,12 +135,129 @@ public:
             else
                 CV_Assert(m1 != NULL);
 
-    #if CV_TRY_AVX2
+#if CV_TRY_AVX2
             if( useAVX2 )
                 j = cv::initUndistortRectifyMapLine_AVX(m1f, m2f, m1, m2,
                                                         matTilt.val, ir, _x, _y, _w, size.width, m1type,
                                                         k1, k2, k3, k4, k5, k6, p1, p2, s1, s2, s3, s4, u0, v0, fx, fy);
-    #endif
+#endif
+#if CV_SIMD_64F
+            const v_float64 v_one = vx_setall_f64(1.0);
+            for (; j <= size.width - 2*v_float64::nlanes; j += 2*v_float64::nlanes, _x += 2*v_float64::nlanes * ir[0], _y += 2*v_float64::nlanes * ir[3], _w += 2*v_float64::nlanes * ir[6])
+            {
+                v_float64 m_0, m_1, m_2, m_3;
+                m_2 = v_one / (vx_setall_f64(_w) + vx_load(s_w));
+                m_3 = v_one / (vx_setall_f64(_w) + vx_load(s_w + v_float64::nlanes));
+                m_0 = vx_setall_f64(_x); m_1 = vx_setall_f64(_y);
+                v_float64 x_0 = (m_0 + vx_load(s_x)) * m_2;
+                v_float64 x_1 = (m_0 + vx_load(s_x + v_float64::nlanes)) * m_3;
+                v_float64 y_0 = (m_1 + vx_load(s_y)) * m_2;
+                v_float64 y_1 = (m_1 + vx_load(s_y + v_float64::nlanes)) * m_3;
+
+                v_float64 xd_0 = x_0 * x_0;
+                v_float64 yd_0 = y_0 * y_0;
+                v_float64 xd_1 = x_1 * x_1;
+                v_float64 yd_1 = y_1 * y_1;
+
+                v_float64 r2_0 = xd_0 + yd_0;
+                v_float64 r2_1 = xd_1 + yd_1;
+
+                m_1 = vx_setall_f64(k3);
+                m_2 = vx_setall_f64(k2);
+                m_3 = vx_setall_f64(k1);
+                m_0 = v_muladd(v_muladd(v_muladd(m_1, r2_0, m_2), r2_0, m_3), r2_0, v_one);
+                m_1 = v_muladd(v_muladd(v_muladd(m_1, r2_1, m_2), r2_1, m_3), r2_1, v_one);
+                m_3 = vx_setall_f64(k6);
+                m_2 = vx_setall_f64(k5);
+                m_0 /= v_muladd(v_muladd(v_muladd(m_3, r2_0, m_2), r2_0, vx_setall_f64(k4)), r2_0, v_one);
+                m_1 /= v_muladd(v_muladd(v_muladd(m_3, r2_1, m_2), r2_1, vx_setall_f64(k4)), r2_1, v_one);
+                x_0 *= m_0; y_0 *= m_0; x_1 *= m_1; y_1 *= m_1;
+
+                m_0 = vx_setall_f64(p1);
+                m_1 = vx_setall_f64(p2);
+                m_2 = vx_setall_f64(2.0);
+                xd_0 = v_muladd(v_muladd(m_2, xd_0, r2_0), m_1, x_0);
+                yd_0 = v_muladd(v_muladd(m_2, yd_0, r2_0), m_0, y_0);
+                xd_1 = v_muladd(v_muladd(m_2, xd_1, r2_1), m_1, x_1);
+                yd_1 = v_muladd(v_muladd(m_2, yd_1, r2_1), m_0, y_1);
+
+                m_0 *= m_2; m_1 *= m_2;
+                m_2 = x_0 * y_0;
+                m_3 = x_1 * y_1;
+                xd_0 = v_muladd(m_0, m_2, xd_0);
+                yd_0 = v_muladd(m_1, m_2, yd_0);
+                xd_1 = v_muladd(m_0, m_3, xd_1);
+                yd_1 = v_muladd(m_1, m_3, yd_1);
+
+                m_0 = r2_0 * r2_0;
+                m_1 = r2_1 * r2_1;
+                m_2 = vx_setall_f64(s2);
+                m_3 = vx_setall_f64(s1);
+                xd_0 = v_muladd(m_3, r2_0, v_muladd(m_2, m_0, xd_0));
+                xd_1 = v_muladd(m_3, r2_1, v_muladd(m_2, m_1, xd_1));
+                m_2 = vx_setall_f64(s4);
+                m_3 = vx_setall_f64(s3);
+                yd_0 = v_muladd(m_3, r2_0, v_muladd(m_2, m_0, yd_0));
+                yd_1 = v_muladd(m_3, r2_1, v_muladd(m_2, m_1, yd_1));
+
+                m_0 = vx_setall_f64(matTilt.val[0]);
+                m_1 = vx_setall_f64(matTilt.val[1]);
+                m_2 = vx_setall_f64(matTilt.val[2]);
+                x_0 = v_muladd(m_0, xd_0, v_muladd(m_1, yd_0, m_2));
+                x_1 = v_muladd(m_0, xd_1, v_muladd(m_1, yd_1, m_2));
+                m_0 = vx_setall_f64(matTilt.val[3]);
+                m_1 = vx_setall_f64(matTilt.val[4]);
+                m_2 = vx_setall_f64(matTilt.val[5]);
+                y_0 = v_muladd(m_0, xd_0, v_muladd(m_1, yd_0, m_2));
+                y_1 = v_muladd(m_0, xd_1, v_muladd(m_1, yd_1, m_2));
+                m_0 = vx_setall_f64(matTilt.val[6]);
+                m_1 = vx_setall_f64(matTilt.val[7]);
+                m_2 = vx_setall_f64(matTilt.val[8]);
+                r2_0 = v_muladd(m_0, xd_0, v_muladd(m_1, yd_0, m_2));
+                r2_1 = v_muladd(m_0, xd_1, v_muladd(m_1, yd_1, m_2));
+                m_0 = vx_setzero_f64();
+                r2_0 = v_select(r2_0 == m_0, v_one, v_one / r2_0);
+                r2_1 = v_select(r2_1 == m_0, v_one, v_one / r2_1);
+
+                m_0 = vx_setall_f64(fx);
+                m_1 = vx_setall_f64(u0);
+                m_2 = vx_setall_f64(fy);
+                m_3 = vx_setall_f64(v0);
+                x_0 = v_muladd(m_0 * r2_0, x_0, m_1);
+                y_0 = v_muladd(m_2 * r2_0, y_0, m_3);
+                x_1 = v_muladd(m_0 * r2_1, x_1, m_1);
+                y_1 = v_muladd(m_2 * r2_1, y_1, m_3);
+
+                if (m1type == CV_32FC1)
+                {
+                    v_store(&m1f[j], v_cvt_f32(x_0, x_1));
+                    v_store(&m2f[j], v_cvt_f32(y_0, y_1));
+                }
+                else if (m1type == CV_32FC2)
+                {
+                    v_float32 mf0, mf1;
+                    v_zip(v_cvt_f32(x_0, x_1), v_cvt_f32(y_0, y_1), mf0, mf1);
+                    v_store(&m1f[j * 2], mf0);
+                    v_store(&m1f[j * 2 + v_float32::nlanes], mf1);
+                }
+                else // m1type == CV_16SC2
+                {
+                    m_0 = vx_setall_f64(INTER_TAB_SIZE);
+                    x_0 *= m_0; x_1 *= m_0; y_0 *= m_0; y_1 *= m_0;
+
+                    v_int32 mask = vx_setall_s32(INTER_TAB_SIZE - 1);
+                    v_int32 iu = v_round(x_0, x_1);
+                    v_int32 iv = v_round(y_0, y_1);
+
+                    v_pack_u_store(&m2[j], (iu & mask) + (iv & mask) * vx_setall_s32(INTER_TAB_SIZE));
+                    v_int32 out0, out1;
+                    v_zip(iu >> INTER_BITS, iv >> INTER_BITS, out0, out1);
+                    v_store(&m1[j * 2], v_pack(out0, out1));
+                }
+            }
+
+            vx_cleanup();
+#endif
             for( ; j < size.width; j++, _x += ir[0], _y += ir[3], _w += ir[6] )
             {
                 double w = 1./_w, x = _x*w, y = _y*w;
@@ -191,6 +317,11 @@ private:
     double s4;
 #if CV_TRY_AVX2
     bool useAVX2;
+#endif
+#if CV_SIMD_64F
+    double s_x[2*v_float64::nlanes];
+    double s_y[2*v_float64::nlanes];
+    double s_w[2*v_float64::nlanes];
 #endif
 };
 }