### This pullrequest changes
WUI based implementation to initUndistortRectifyMap