@terfendail 2019-07-09
Performance change for InitUndistortMap::Undistort

| Baseline | Reference time |  PR time | Speedup |
|------------------|----------------|----------|---------|
|SSE2|4.561|2.189|2.08|
|SSE3|4.560|2.195|2.08|
|SSE4_2|4.452|2.069|2.15|
|AVX2|0.930|0.860|1.08|